+++ Example1/statement.js
@@ -1,0 +1,38 @@
+//@ts-check
+
+import createStatementData from './createStatementData.js';
+
+
+
+/**
+ * 
+ * @param {Object} invoice 
+ * @param {Object} plays 
+ * @returns {string}
+ */
+export function statement(invoice, plays) {
+    
+    return renderPlainText(createStatementData(invoice, plays));
+}
+
+
+function renderPlainText(data) {
+    let result = `Statement for ${data.customer}\n`;
+
+    for (let perf of data.performances) {
+        result += ` ${perf.play.name}: ${usd(perf.amount)} (${perf.audience} seats)`;
+    }
+
+    result += `Amount owed is ${usd(data.totalAmount)}\n`;
+    result += `You earned ${data.totalVolumeCredits} credits`;
+
+    return result;
+}
+
+
+function usd(aNumber) {
+    return new Intl.NumberFormat("en-US", {
+        style: "currency", currency: "USD", minimumFractionDigits: 2
+    }).format(aNumber / 100);
+}
+
